pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = "Process batch data."
name = "data_processing"
version = "0.1.0"

[tool.poetry.dependencies]
aiofiles = "^0.6.0"
boto3 = "^1.17.54"
celery = "^5.0.5"
fastapi = "^0.63.0"
mongoengine = "^0.23.0"
python = "^3.8"
python-multipart = "^0.0.5"
pyyaml = "^5.4.1"
redis = "^3.5.3"
uvicorn = "^0.13.4"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}
prospector = "^1.3.1"

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
